fix types and things
