+++ lib/handlers/api.ts
@@ -154,7 +154,7 @@ export class ApiHandler {
             .then(result => {
                 const config = JSON.parse(result.config);
 
-                res.header('Link-Created', result.created.toUTCString());
+                if (result.created) res.header('Link-Created', result.created.toUTCString());
 
                 if (config.content) {
                     const normalizer = new ClientStateNormalizer();
