+++ test/unfurl-tests.ts
@@ -76,10 +76,8 @@ describe('Basic unfurls', () => {
 
         const res = await prom;
         res.metadata.should.deep.equal({
-            ogAuthor: null,
             ogDescription: '',
             ogTitle: 'Compiler Explorer',
-            ogCreated: undefined,
         });
     });
 
@@ -106,11 +104,9 @@ describe('Basic unfurls', () => {
 
         const res = await prom;
         res.metadata.should.deep.equal({
-            ogAuthor: null,
             ogDescription:
                 '\ntemplate&lt;typename T&gt;\nconcept TheSameAndAddable = requires(T a, T b) {\n    {a+b} -&gt; T;\n};\n\ntemplate&lt;TheSameAndAddable T&gt;\nT sum(T x, T y) {\n    return x + y;\n}\n\n#include &lt;string&gt;\n\nint main() {\n    int z = 0;\n    int w;\n\n    return sum(z, w);\n}\n',
             ogTitle: 'Compiler Explorer - C++',
-            ogCreated: undefined,
         });
     });
 
@@ -137,10 +133,8 @@ describe('Basic unfurls', () => {
 
         const res = await prom;
         res.metadata.should.deep.equal({
-            ogAuthor: null,
             ogDescription: 'project(hello)\n\nadd_executable(output.s\n    example.cpp\n    square.cpp)\n',
             ogTitle: 'Compiler Explorer - C++',
-            ogCreated: undefined,
         });
     });
 });
